Vibe/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework
dj_database_url
gunicorn
django-cors-headers
requests